Definitive 'Doors' script update
We are too much busy rigth now, somethings went wrong in production and because of that I will not describe change by change.
Summing up, it's possible to rotate any type of doors in most of type of rotations doesn't matter the father object rotation.
Just put the values you want and be happy with the smooth rotation!

PS: The correct rotate doors script name will be changed in the next commit to not get conflicts because the archive name. By now the name is 'DoorsTest2'.