FMOD Major Update
Setted up the project to work with FMOD. The major changes in scripts it was in the 'Dialogue' (changed) and a new script to play the sound 'Playsound' that have the ISound interface (need to improve later).

The following scenes like: '1_Introcucao' and '2_Sonho 1' was added the dubbing feature.

Also, I still seeking about getLength(); to optimize the general process of implementation, for now, using the '_countProvisorio' float variable to set the dialogue text time in sceen to follow the dubbing time.